backends/aafnoghar_backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
pillow==11.2.1
channels==4.0.0
daphne==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
